Add partial refactor of loss module
