+++ src/nn/loss.rs
@@ -27,14 +27,17 @@
 //!
 //! * [`kldiv_loss`] -  Measures the Kullback-Leibler divergence between the target and the input.
 use super::{
-    variable::{expect_tensor, expect_tensor_mut, node::Overwrite},
-    Backward, Data, Forward, Gradient, Tensor, Var, VarDiff,
+    variable::node::{
+        backward::{
+            BCELossBackward, BCEWithLogitsLossBackward, KLDivLossBackward, MAELossBackward,
+            MSELossBackward, NLLLossBackward,
+        },
+        forward::{BCELoss, BCEWithLogitsLoss, KLDivLoss, MAELoss, MSELoss, NLLLoss},
+        Overwrite,
+    },
+    Data, Gradient, Var, VarDiff,
 };
-use ndarray::{Axis, Dimension, IntoDimension, Ix1, Zip};
-use std::{
-    cell::{Cell, Ref, RefCell, RefMut},
-    rc::Rc,
-};
+use ndarray::Dimension;
 
 /// Specifies the reduction to apply to the *loss* output.
 #[derive(Clone)]
@@ -44,218 +47,6 @@ pub enum Reduction {
     /// The sum of the output will be divided by the batch size for the [`kldiv_loss`] and the
     /// [`nll_loss`]. For all other losses the output will be divided by the number of elements.
     Mean,
-}
-
-// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Mean Square Error Loss ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#[allow(clippy::upper_case_acronyms)]
-pub struct MSELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    input: Rc<T>,
-    target: Rc<U>,
-    data: RefCell<Tensor<Ix1>>,
-    reduction: Reduction,
-    computed: Cell<bool>,
-}
-
-impl<T, U> MSELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    pub(crate) fn new(input: Rc<T>, target: Rc<U>, reduction: Reduction) -> Self {
-        Self {
-            input,
-            target,
-            data: RefCell::new(Tensor::zeros(1)),
-            reduction,
-            computed: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U> Data for MSELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    type Dim = Ix1;
-
-    fn data(&self) -> Ref<Tensor<Self::Dim>> {
-        self.data.borrow()
-    }
-
-    fn data_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        self.data.borrow_mut()
-    }
-}
-
-impl<T, U> Forward for MSELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    fn forward(&self) {
-        if self.was_computed() {
-            return;
-        }
-
-        self.computed.set(true);
-        let (mut loss_data, input_data, target_data) = {
-            (
-                self.data.borrow_mut(),
-                self.input.data(),
-                self.target.data(),
-            )
-        };
-        loss_data[0] = {
-            let total_loss = Zip::from(&*input_data)
-                .and(&*target_data)
-                .fold(0.0, |loss, input, target| loss + (input - target).powi(2));
-            match self.reduction {
-                Reduction::Mean => total_loss / input_data.len() as f32,
-                Reduction::Sum => total_loss,
-            }
-        };
-    }
-
-    fn was_computed(&self) -> bool {
-        self.computed.get()
-    }
-
-    fn reset_computation(&self) {
-        self.computed.set(false);
-    }
-}
-
-#[allow(clippy::upper_case_acronyms)]
-pub struct MSELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = U::Dim>,
-{
-    gradient: RefCell<Option<Tensor<Ix1>>>,
-    overwrite: Cell<bool>,
-    diff_input: Rc<T>,
-    input: Rc<U>,
-    target: Rc<V>,
-    reduction: Reduction,
-}
-
-impl<T, U, V> MSELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = U::Dim>,
-{
-    pub(crate) fn new(
-        diff_input: Rc<T>,
-        input: Rc<U>,
-        target: Rc<V>,
-        reduction: Reduction,
-    ) -> Self {
-        Self {
-            diff_input,
-            input,
-            target,
-            gradient: RefCell::new(Some(Tensor::zeros(1))),
-            reduction,
-            overwrite: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U, V> Gradient for MSELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = U::Dim>,
-{
-    type Dim = Ix1;
-
-    fn gradient(&self) -> Ref<Tensor<Self::Dim>> {
-        expect_tensor(&self.gradient)
-    }
-
-    fn gradient_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        expect_tensor_mut(&self.gradient)
-    }
-}
-
-impl<T, U, V> Overwrite for MSELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    fn can_overwrite(&self) -> bool {
-        self.overwrite.get()
-    }
-
-    fn set_overwrite(&self, state: bool) {
-        self.overwrite.set(state);
-    }
-}
-
-impl<T, U, V> Backward for MSELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    fn backward(&self) {
-        let (mut operand_gradient, gradient, input_data, target_data) = {
-            (
-                self.diff_input.gradient_mut(),
-                self.gradient(),
-                self.input.data(),
-                self.target.data(),
-            )
-        };
-
-        let zip = Zip::from(&mut *operand_gradient)
-            .and_broadcast(&*gradient)
-            .and(&*input_data)
-            .and(&*target_data);
-        match self.reduction {
-            Reduction::Mean => {
-                let n = input_data.len() as f32;
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        *op_grad = (2.0 * (input - target)) * grad / n
-                    });
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        *op_grad += (2.0 * (input - target)) * grad / n
-                    });
-                }
-            }
-            Reduction::Sum => {
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        *op_grad = (2.0 * (input - target)) * grad
-                    });
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        *op_grad += (2.0 * (input - target)) * grad
-                    });
-                }
-            }
-        }
-    }
-
-    fn no_grad(&self) {
-        *self.gradient.borrow_mut() = None;
-    }
-
-    fn with_grad(&self) {
-        *self.gradient.borrow_mut() = Some(Tensor::zeros(1));
-    }
 }
 
 /// Computes the **mean squared error** *(squared L2 norm)* between each element in the input x
@@ -288,231 +79,6 @@ where
     VarDiff::from(backward_node, input.past, var)
 }
 
-// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Mean Absolute Error Loss ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#[allow(clippy::upper_case_acronyms)]
-pub struct MAELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    input: Rc<T>,
-    target: Rc<U>,
-    data: RefCell<Tensor<Ix1>>,
-    reduction: Reduction,
-    computed: Cell<bool>,
-}
-
-impl<T, U> MAELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    pub(crate) fn new(input: Rc<T>, target: Rc<U>, reduction: Reduction) -> Self {
-        Self {
-            input,
-            target,
-            data: RefCell::new(Tensor::zeros(1)),
-            reduction,
-            computed: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U> Data for MAELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    type Dim = Ix1;
-
-    fn data(&self) -> Ref<Tensor<Self::Dim>> {
-        self.data.borrow()
-    }
-
-    fn data_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        self.data.borrow_mut()
-    }
-}
-
-impl<T, U> Forward for MAELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    fn forward(&self) {
-        if self.was_computed() {
-            return;
-        }
-
-        self.computed.set(true);
-        let (mut loss_data, input_data, target_data) = {
-            (
-                self.data.borrow_mut(),
-                self.input.data(),
-                self.target.data(),
-            )
-        };
-        loss_data[0] = {
-            let total_loss = Zip::from(&*input_data)
-                .and(&*target_data)
-                .fold(0.0, |loss, input, target| loss + (input - target).abs());
-            match self.reduction {
-                Reduction::Mean => total_loss / input_data.len() as f32,
-                Reduction::Sum => total_loss,
-            }
-        };
-    }
-
-    fn was_computed(&self) -> bool {
-        self.computed.get()
-    }
-
-    fn reset_computation(&self) {
-        self.computed.set(false);
-    }
-}
-
-#[allow(clippy::upper_case_acronyms)]
-pub struct MAELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    gradient: RefCell<Option<Tensor<Ix1>>>,
-    overwrite: Cell<bool>,
-    diff_input: Rc<T>,
-    input: Rc<U>,
-    target: Rc<V>,
-    reduction: Reduction,
-}
-
-impl<T, U, V> MAELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    pub(crate) fn new(
-        diff_input: Rc<T>,
-        input: Rc<U>,
-        target: Rc<V>,
-        reduction: Reduction,
-    ) -> Self {
-        Self {
-            diff_input,
-            input,
-            target,
-            gradient: RefCell::new(Some(Tensor::zeros(1))),
-            reduction,
-            overwrite: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U, V> Gradient for MAELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    type Dim = Ix1;
-
-    fn gradient(&self) -> Ref<Tensor<Self::Dim>> {
-        expect_tensor(&self.gradient)
-    }
-
-    fn gradient_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        expect_tensor_mut(&self.gradient)
-    }
-}
-
-impl<T, U, V> Overwrite for MAELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    fn can_overwrite(&self) -> bool {
-        self.overwrite.get()
-    }
-
-    fn set_overwrite(&self, state: bool) {
-        self.overwrite.set(state);
-    }
-}
-
-impl<T, U, V> Backward for MAELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    fn backward(&self) {
-        let (mut operand_gradient, gradient, input_data, target_data) = {
-            (
-                self.diff_input.gradient_mut(),
-                self.gradient(),
-                self.input.data(),
-                self.target.data(),
-            )
-        };
-
-        let zip = Zip::from(&mut *operand_gradient)
-            .and_broadcast(&*gradient)
-            .and(&*input_data)
-            .and(&*target_data);
-
-        match self.reduction {
-            Reduction::Mean => {
-                let n = input_data.len() as f32;
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        let diff = input - target;
-                        *op_grad = if diff != 0. {
-                            diff.signum() * grad / n
-                        } else {
-                            0.
-                        }
-                    });
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        let diff = input - target;
-                        *op_grad += if diff != 0. {
-                            diff.signum() * grad / n
-                        } else {
-                            0.
-                        }
-                    });
-                }
-            }
-            Reduction::Sum => {
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        let diff = input - target;
-                        *op_grad = if diff != 0. { diff.signum() * grad } else { 0. }
-                    });
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        let diff = input - target;
-                        *op_grad += if diff != 0. { diff.signum() * grad } else { 0. }
-                    });
-                }
-            }
-        }
-    }
-
-    fn no_grad(&self) {
-        *self.gradient.borrow_mut() = None;
-    }
-
-    fn with_grad(&self) {
-        *self.gradient.borrow_mut() = Some(Tensor::zeros(1));
-    }
-}
-
 /// Computes the **mean absolute error** *(MAE)* between each element in the input x and target y.
 ///
 /// ```text
@@ -540,231 +106,6 @@ where
 
     let backward_node = MAELossBackward::new(input.node, input.var.node, target.node, reduction);
     VarDiff::from(backward_node, input.past, var)
-}
-
-// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Binary Cross Entropy Loss ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-#[allow(clippy::upper_case_acronyms)]
-pub struct BCELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    input: Rc<T>,
-    target: Rc<U>,
-    data: RefCell<Tensor<Ix1>>,
-    reduction: Reduction,
-    computed: Cell<bool>,
-}
-
-impl<T, U> BCELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    pub(crate) fn new(input: Rc<T>, target: Rc<U>, reduction: Reduction) -> Self {
-        Self {
-            input,
-            target,
-            data: RefCell::new(Tensor::zeros(1)),
-            reduction,
-            computed: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U> Data for BCELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    type Dim = Ix1;
-
-    fn data(&self) -> Ref<Tensor<Self::Dim>> {
-        self.data.borrow()
-    }
-
-    fn data_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        self.data.borrow_mut()
-    }
-}
-
-impl<T, U> Forward for BCELoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    fn forward(&self) {
-        if self.was_computed() {
-            return;
-        }
-
-        self.computed.set(true);
-        let (mut loss_data, input_data, target_data) = {
-            (
-                self.data.borrow_mut(),
-                self.input.data(),
-                self.target.data(),
-            )
-        };
-        const MIN_LOG: f32 = -100.;
-        loss_data[0] = {
-            let total_loss =
-                Zip::from(&*input_data)
-                    .and(&*target_data)
-                    .fold(0.0, |loss, input, target| {
-                        loss + (target * input.ln().clamp(MIN_LOG, std::f32::MAX)
-                            + (1. - target) * (1. - input).ln().clamp(MIN_LOG, std::f32::MAX))
-                    });
-            match self.reduction {
-                Reduction::Mean => -total_loss / input_data.len() as f32,
-                Reduction::Sum => -total_loss,
-            }
-        };
-    }
-
-    fn was_computed(&self) -> bool {
-        self.computed.get()
-    }
-
-    fn reset_computation(&self) {
-        self.computed.set(false);
-    }
-}
-
-#[allow(clippy::upper_case_acronyms)]
-pub struct BCELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    gradient: RefCell<Option<Tensor<Ix1>>>,
-    overwrite: Cell<bool>,
-    diff_input: Rc<T>,
-    input: Rc<U>,
-    target: Rc<V>,
-    reduction: Reduction,
-}
-
-impl<T, U, V> BCELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    pub(crate) fn new(
-        diff_input: Rc<T>,
-        input: Rc<U>,
-        target: Rc<V>,
-        reduction: Reduction,
-    ) -> Self {
-        Self {
-            diff_input,
-            input,
-            target,
-            gradient: RefCell::new(Some(Tensor::zeros(1))),
-            reduction,
-            overwrite: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U, V> Gradient for BCELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    type Dim = Ix1;
-
-    fn gradient(&self) -> Ref<Tensor<Self::Dim>> {
-        expect_tensor(&self.gradient)
-    }
-
-    fn gradient_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        expect_tensor_mut(&self.gradient)
-    }
-}
-
-impl<T, U, V> Overwrite for BCELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    fn can_overwrite(&self) -> bool {
-        self.overwrite.get()
-    }
-
-    fn set_overwrite(&self, state: bool) {
-        self.overwrite.set(state);
-    }
-}
-
-impl<T, U, V> Backward for BCELossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = U::Dim>,
-{
-    fn backward(&self) {
-        let (mut operand_gradient, gradient, input_data, target_data) = {
-            (
-                self.diff_input.gradient_mut(),
-                self.gradient(),
-                self.input.data(),
-                self.target.data(),
-            )
-        };
-
-        let zip = Zip::from(&mut *operand_gradient)
-            .and_broadcast(&*gradient)
-            .and(&*input_data)
-            .and(&*target_data);
-
-        match self.reduction {
-            Reduction::Mean => {
-                let n = input_data.len() as f32;
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        *op_grad = (input - target) / ((1. - input) * input).max(std::f32::EPSILON)
-                            * grad
-                            / n
-                    });
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        *op_grad += (input - target) / ((1. - input) * input).max(std::f32::EPSILON)
-                            * grad
-                            / n
-                    });
-                }
-            }
-            Reduction::Sum => {
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        *op_grad =
-                            (input - target) / ((1. - input) * input).max(std::f32::EPSILON) * grad
-                    });
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        *op_grad +=
-                            (input - target) / ((1. - input) * input).max(std::f32::EPSILON) * grad
-                    });
-                }
-            }
-        }
-    }
-
-    fn no_grad(&self) {
-        *self.gradient.borrow_mut() = None;
-    }
-
-    fn with_grad(&self) {
-        *self.gradient.borrow_mut() = Some(Tensor::zeros(1));
-    }
 }
 
 /// Computes the **binary cross entropy** between the target y and input x.
@@ -803,229 +144,6 @@ where
     VarDiff::from(backward_node, input.past, var)
 }
 
-// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Binary Cross Entropy With Logits Loss ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#[allow(clippy::upper_case_acronyms)]
-pub struct BCEWithLogitsLoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    input: Rc<T>,
-    target: Rc<U>,
-    data: RefCell<Tensor<Ix1>>,
-    reduction: Reduction,
-    state: Cell<bool>,
-}
-
-impl<T, U> BCEWithLogitsLoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    pub(crate) fn new(input: Rc<T>, target: Rc<U>, reduction: Reduction) -> Self {
-        Self {
-            input,
-            target,
-            data: RefCell::new(Tensor::zeros(1)),
-            reduction,
-            state: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U> Data for BCEWithLogitsLoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    type Dim = Ix1;
-
-    fn data(&self) -> Ref<Tensor<Self::Dim>> {
-        self.data.borrow()
-    }
-
-    fn data_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        self.data.borrow_mut()
-    }
-}
-
-impl<T, U> Forward for BCEWithLogitsLoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    fn forward(&self) {
-        if self.was_computed() {
-            return;
-        }
-
-        self.state.set(true);
-        let (mut loss_data, input_data, target_data) = {
-            (
-                self.data.borrow_mut(),
-                self.input.data(),
-                self.target.data(),
-            )
-        };
-        loss_data[0] = {
-            let total_loss =
-                Zip::from(&*input_data)
-                    .and(&*target_data)
-                    .fold(0.0, |loss, input, target| {
-                        let max = (-input).max(0.);
-                        loss + (1. - target) * input
-                            + max
-                            + ((-max).exp() + (-input - max).exp()).ln()
-                    });
-            match self.reduction {
-                Reduction::Mean => total_loss / input_data.len() as f32,
-                Reduction::Sum => total_loss,
-            }
-        };
-    }
-
-    fn was_computed(&self) -> bool {
-        self.state.get()
-    }
-
-    fn reset_computation(&self) {
-        self.state.set(false);
-    }
-}
-
-#[allow(clippy::upper_case_acronyms)]
-pub struct BCEWithLogitsLossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    gradient: RefCell<Option<Tensor<Ix1>>>,
-    overwrite: Cell<bool>,
-    diff_input: Rc<T>,
-    input: Rc<U>,
-    target: Rc<V>,
-    reduction: Reduction,
-}
-
-impl<T, U, V> BCEWithLogitsLossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    pub(crate) fn new(
-        diff_input: Rc<T>,
-        input: Rc<U>,
-        target: Rc<V>,
-        reduction: Reduction,
-    ) -> Self {
-        Self {
-            diff_input,
-            input,
-            target,
-            gradient: RefCell::new(Some(Tensor::zeros(1))),
-            reduction,
-            overwrite: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U, V> Gradient for BCEWithLogitsLossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    type Dim = Ix1;
-
-    fn gradient(&self) -> Ref<Tensor<Self::Dim>> {
-        expect_tensor(&self.gradient)
-    }
-
-    fn gradient_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        expect_tensor_mut(&self.gradient)
-    }
-}
-
-impl<T, U, V> Overwrite for BCEWithLogitsLossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    fn can_overwrite(&self) -> bool {
-        self.overwrite.get()
-    }
-
-    fn set_overwrite(&self, state: bool) {
-        self.overwrite.set(state);
-    }
-}
-
-impl<T, U, V> Backward for BCEWithLogitsLossBackward<T, U, V>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-    V: Data<Dim = T::Dim>,
-{
-    fn backward(&self) {
-        let (mut operand_gradient, gradient, input_data, target_data) = {
-            (
-                self.diff_input.gradient_mut(),
-                self.gradient(),
-                self.input.data(),
-                self.target.data(),
-            )
-        };
-
-        let zip = Zip::from(&mut *operand_gradient)
-            .and_broadcast(&*gradient)
-            .and(&*input_data)
-            .and(&*target_data);
-
-        match self.reduction {
-            Reduction::Mean => {
-                let n = input_data.len() as f32;
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        let input_sigmoid = 1.0 / (1.0 + (-input).exp());
-                        *op_grad = (input_sigmoid - target) * grad / n
-                    });
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        let input_sigmoid = 1.0 / (1.0 + (-input).exp());
-                        *op_grad += (input_sigmoid - target) * grad / n
-                    });
-                }
-            }
-            Reduction::Sum => {
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        let input_sigmoid = 1.0 / (1.0 + (-input).exp());
-                        *op_grad = (input_sigmoid - target) * grad
-                    });
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|op_grad, grad, input, target| {
-                        let input_sigmoid = 1.0 / (1.0 + (-input).exp());
-                        *op_grad += (input_sigmoid - target) * grad
-                    });
-                }
-            }
-        }
-    }
-
-    fn no_grad(&self) {
-        *self.gradient.borrow_mut() = None;
-    }
-
-    fn with_grad(&self) {
-        *self.gradient.borrow_mut() = Some(Tensor::zeros(1));
-    }
-}
-
 /// Computes the **binary cross entropy with logits** between the target y and input x.
 ///
 /// ```text
@@ -1060,234 +178,6 @@ where
     let backward_node =
         BCEWithLogitsLossBackward::new(input.node, input.var.node, target.node, reduction);
     VarDiff::from(backward_node, input.past, var)
-}
-
-// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ NLLLoss ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#[allow(clippy::upper_case_acronyms)]
-pub struct NLLLoss<T, U>
-where
-    T: Data<Dim = <U::Dim as Dimension>::Larger>,
-    T::Dim: Copy,
-    U: Data,
-{
-    input: Rc<T>,
-    target: Rc<U>,
-    data: RefCell<Tensor<Ix1>>,
-    reduction: Reduction,
-    computed: Cell<bool>,
-}
-
-impl<T, U> NLLLoss<T, U>
-where
-    T: Data<Dim = <U::Dim as Dimension>::Larger>,
-    T::Dim: Copy,
-    U: Data,
-{
-    pub(crate) fn new(input: Rc<T>, target: Rc<U>, reduction: Reduction) -> Self {
-        Self {
-            input,
-            target,
-            data: RefCell::new(Tensor::zeros(1)),
-            reduction,
-            computed: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U> Data for NLLLoss<T, U>
-where
-    T: Data<Dim = <U::Dim as Dimension>::Larger>,
-    T::Dim: Copy,
-    U: Data,
-{
-    type Dim = Ix1;
-
-    fn data(&self) -> Ref<Tensor<Self::Dim>> {
-        self.data.borrow()
-    }
-
-    fn data_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        self.data.borrow_mut()
-    }
-}
-
-impl<T, U> Forward for NLLLoss<T, U>
-where
-    T: Data<Dim = <U::Dim as Dimension>::Larger>,
-    T::Dim: Copy,
-    U: Data,
-{
-    fn forward(&self) {
-        if self.was_computed() {
-            return;
-        }
-        self.computed.set(true);
-        let (mut loss_data, input_data, target_data) = {
-            (
-                self.data.borrow_mut(),
-                self.input.data(),
-                self.target.data(),
-            )
-        };
-        loss_data[0] = {
-            let total_loss = Zip::indexed(&*input_data)
-                .and_broadcast(&target_data.view().insert_axis(Axis(1)))
-                .fold(0.0, |loss, idx, log, target| {
-                    if idx.into_dimension()[1] == *target as usize {
-                        loss + log
-                    } else {
-                        loss + 0.
-                    }
-                });
-            match self.reduction {
-                Reduction::Mean => -total_loss / input_data.len_of(Axis(0)) as f32,
-                Reduction::Sum => -total_loss,
-            }
-        };
-    }
-
-    fn was_computed(&self) -> bool {
-        self.computed.get()
-    }
-
-    fn reset_computation(&self) {
-        self.computed.set(false);
-    }
-}
-
-#[allow(clippy::upper_case_acronyms)]
-pub struct NLLLossBackward<T, U>
-where
-    T: Gradient<Dim = <U::Dim as Dimension>::Larger> + Overwrite,
-    U: Data,
-    T::Dim: Copy,
-{
-    diff_input: Rc<T>,
-    target: Rc<U>,
-    gradient: RefCell<Option<Tensor<Ix1>>>,
-    reduction: Reduction,
-    overwrite: Cell<bool>,
-}
-
-impl<T, U> NLLLossBackward<T, U>
-where
-    T: Gradient<Dim = <U::Dim as Dimension>::Larger> + Overwrite,
-    U: Data,
-    T::Dim: Copy,
-{
-    pub(crate) fn new(diff_input: Rc<T>, target: Rc<U>, reduction: Reduction) -> Self {
-        Self {
-            diff_input,
-            target,
-            gradient: RefCell::new(Some(Tensor::zeros(1))),
-            reduction,
-            overwrite: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U> Gradient for NLLLossBackward<T, U>
-where
-    T: Gradient<Dim = <U::Dim as Dimension>::Larger> + Overwrite,
-    U: Data,
-    T::Dim: Copy,
-{
-    type Dim = Ix1;
-
-    fn gradient(&self) -> Ref<Tensor<Self::Dim>> {
-        expect_tensor(&self.gradient)
-    }
-
-    fn gradient_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        expect_tensor_mut(&self.gradient)
-    }
-}
-
-impl<T, U> Overwrite for NLLLossBackward<T, U>
-where
-    T: Gradient<Dim = <U::Dim as Dimension>::Larger> + Overwrite,
-    U: Data,
-    T::Dim: Copy,
-{
-    fn can_overwrite(&self) -> bool {
-        self.overwrite.get()
-    }
-
-    fn set_overwrite(&self, state: bool) {
-        self.overwrite.set(state);
-    }
-}
-
-impl<T, U> Backward for NLLLossBackward<T, U>
-where
-    T: Gradient<Dim = <U::Dim as Dimension>::Larger> + Overwrite,
-    U: Data,
-    T::Dim: Copy,
-{
-    fn backward(&self) {
-        let (mut operand_gradient, gradient, target_data) = {
-            (
-                self.diff_input.gradient_mut(),
-                self.gradient(),
-                self.target.data(),
-            )
-        };
-        let zip = Zip::indexed(&mut *operand_gradient)
-            .and_broadcast(&*gradient)
-            .and_broadcast(target_data.view().insert_axis(Axis(1)));
-
-        match self.reduction {
-            Reduction::Mean => {
-                let n = target_data.len() as f32;
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|idx, op_grad, grad, target| {
-                        if idx.into_dimension().last_elem() == *target as usize {
-                            *op_grad = grad * -1. / n
-                        } else {
-                            *op_grad = 0.;
-                        }
-                    });
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|idx, op_grad, grad, target| {
-                        if idx.into_dimension().last_elem() == *target as usize {
-                            *op_grad += grad * -1. / n
-                        } else {
-                            *op_grad += 0.;
-                        }
-                    });
-                }
-            }
-            Reduction::Sum => {
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|idx, op_grad, grad, target| {
-                        if idx.into_dimension().last_elem() == *target as usize {
-                            *op_grad = grad * -1.
-                        } else {
-                            *op_grad = 0.
-                        }
-                    });
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|idx, op_grad, grad, target| {
-                        if idx.into_dimension().last_elem() == *target as usize {
-                            *op_grad += grad * -1.
-                        } else {
-                            *op_grad += 0.
-                        }
-                    });
-                }
-            }
-        }
-    }
-
-    fn no_grad(&self) {
-        *self.gradient.borrow_mut() = None;
-    }
-
-    fn with_grad(&self) {
-        *self.gradient.borrow_mut() = Some(Tensor::zeros(1));
-    }
 }
 
 /// Computes the **negative log likelihood** between the target y and input x.
@@ -1336,203 +226,6 @@ where
     VarDiff::from(backward_node, input.past, var)
 }
 
-// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ KLDivLoss ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#[allow(clippy::upper_case_acronyms)]
-pub struct KLDivLoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    input: Rc<T>,
-    target: Rc<U>,
-    data: RefCell<Tensor<Ix1>>,
-    reduction: Reduction,
-    computed: Cell<bool>,
-}
-
-impl<T, U> KLDivLoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    pub(crate) fn new(input: Rc<T>, target: Rc<U>, reduction: Reduction) -> Self {
-        Self {
-            input,
-            target,
-            data: RefCell::new(Tensor::zeros(1)),
-            reduction,
-            computed: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U> Data for KLDivLoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    type Dim = Ix1;
-
-    fn data(&self) -> Ref<Tensor<Self::Dim>> {
-        self.data.borrow()
-    }
-
-    fn data_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        self.data.borrow_mut()
-    }
-}
-
-impl<T, U> Forward for KLDivLoss<T, U>
-where
-    T: Data,
-    U: Data<Dim = T::Dim>,
-{
-    fn forward(&self) {
-        if self.was_computed() {
-            return;
-        }
-        self.computed.set(true);
-        let (mut loss_data, input_data, target_data) = {
-            (
-                self.data.borrow_mut(),
-                self.input.data(),
-                self.target.data(),
-            )
-        };
-        loss_data[0] = {
-            let total_loss =
-                Zip::from(&*input_data)
-                    .and(&*target_data)
-                    .fold(0.0, |loss, log, target| {
-                        if *target > 0. {
-                            loss + target * (target.ln() - log)
-                        } else {
-                            loss + 0.
-                        }
-                    });
-            match self.reduction {
-                Reduction::Mean => total_loss / input_data.len_of(Axis(0)) as f32,
-                Reduction::Sum => total_loss,
-            }
-        };
-    }
-
-    fn was_computed(&self) -> bool {
-        self.computed.get()
-    }
-
-    fn reset_computation(&self) {
-        self.computed.set(false);
-    }
-}
-
-// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ KLDivLossBackward ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#[allow(clippy::upper_case_acronyms)]
-pub struct KLDivLossBackward<T, U>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-{
-    diff_input: Rc<T>,
-    target: Rc<U>,
-    gradient: RefCell<Option<Tensor<Ix1>>>,
-    reduction: Reduction,
-    overwrite: Cell<bool>,
-}
-
-impl<T, U> KLDivLossBackward<T, U>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-{
-    pub(crate) fn new(diff_input: Rc<T>, target: Rc<U>, reduction: Reduction) -> Self {
-        Self {
-            diff_input,
-            target,
-            gradient: RefCell::new(Some(Tensor::zeros(1))),
-            reduction,
-            overwrite: Cell::new(false),
-        }
-    }
-}
-
-impl<T, U> Gradient for KLDivLossBackward<T, U>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-{
-    type Dim = Ix1;
-
-    fn gradient(&self) -> Ref<Tensor<Self::Dim>> {
-        expect_tensor(&self.gradient)
-    }
-
-    fn gradient_mut(&self) -> RefMut<Tensor<Self::Dim>> {
-        expect_tensor_mut(&self.gradient)
-    }
-}
-
-impl<T, U> Overwrite for KLDivLossBackward<T, U>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-{
-    fn can_overwrite(&self) -> bool {
-        self.overwrite.get()
-    }
-
-    fn set_overwrite(&self, state: bool) {
-        self.overwrite.set(state);
-    }
-}
-
-impl<T, U> Backward for KLDivLossBackward<T, U>
-where
-    T: Gradient + Overwrite,
-    U: Data<Dim = T::Dim>,
-{
-    fn backward(&self) {
-        let (mut operand_gradient, gradient, target_data) = {
-            (
-                self.diff_input.gradient_mut(),
-                self.gradient(),
-                self.target.data(),
-            )
-        };
-        let zip = Zip::from(&mut *operand_gradient)
-            .and_broadcast(&*gradient)
-            .and(&*target_data);
-
-        match self.reduction {
-            Reduction::Mean => {
-                let n = target_data.len_of(Axis(0)) as f32;
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|op_grad, grad, target| *op_grad = -target * grad / n);
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|op_grad, grad, target| *op_grad += -target * grad / n);
-                }
-            }
-            Reduction::Sum => {
-                if self.diff_input.can_overwrite() {
-                    zip.for_each(|op_grad, grad, target| *op_grad = -target * grad);
-                    self.diff_input.set_overwrite(false);
-                } else {
-                    zip.for_each(|op_grad, grad, target| *op_grad += -target * grad);
-                }
-            }
-        }
-    }
-
-    fn no_grad(&self) {
-        *self.gradient.borrow_mut() = None;
-    }
-
-    fn with_grad(&self) {
-        *self.gradient.borrow_mut() = Some(Tensor::zeros(1));
-    }
-}
-
 /// Computes the **Kullback-Leibler** divergence between the target and the input.
 ///
 /// ```text
@@ -1572,11 +265,21 @@ where
     VarDiff::from(backward_node, input.past, var)
 }
 
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ Test ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::{variable::node::Differentiable, Input, InputBackward};
-    use ndarray::{Dimension, StrideShape};
+    use crate::{
+        variable::{
+            node::{Backward, Differentiable, Forward},
+            Tensor,
+        },
+        Input, InputBackward,
+    };
+    use ndarray::{Dimension, StrideShape, Zip};
+    use std::rc::Rc;
 
     const F16_EPSILON: f32 = 9.77e-04;
